+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
  * added `PhpRedisStore`
  * added `CachedStore`
+ * removed optional argument `$cache` from `JsonFileStore::__construct()`
 
 * 1.0.0-beta (2015-01-12)